Rename general error handler to API error handler
