+++ frontend/src/components/Login.js
@@ -14,8 +14,8 @@
 
 import React, { Component } from "react";
 import { Redirect } from "react-router-dom";
-import { handleResponseError } from "./utils/GeneralErrorHandler";
-import { errorRedirect } from "./utils/GeneralErrorHandler";
+import { handleResponseError } from "./utils/APIErrorHandler";
+import { errorRedirect } from "./utils/APIErrorHandler";
 import { loading } from "./utils/Utilities";
 
 class Home extends Component {
